[GOBBLIN-255] Add javadoc and remove unnecessary int casts

+++ gobblin-modules/gobblin-parquet/src/main/java/org/apache/gobblin/writer/ParquetHdfsDataWriter.java
@@ -44,7 +44,7 @@ public class ParquetHdfsDataWriter extends FsDataWriter<Group> {
   public ParquetHdfsDataWriter(ParquetDataWriterBuilder builder, State state)
       throws IOException {
     super(builder, state);
-    this.writer = builder.getWriter(this.blockSize, this.stagingFile);
+    this.writer = builder.getWriter((int) this.blockSize, this.stagingFile);
   }
 
   @Override
